backend/requirements.txt
openai>=1.15
fastapi
uvicorn
python-dotenv
